Worker

+++ calc-worker-dotnet/CalcWorker/Api/JobDTO.cs
@@ -7,7 +7,7 @@ namespace CalcWorker.Api
         public string Version { get; set; }
         public int Id { get; set; }
         public int? Input { get; set; }
-        public int? Output { get; set; }
+        public long? Output { get; set; }
         public DateTime? CalcStartedAt { get; set; }
         public DateTime? FinishedAt { get; set; }
         public string Status { get; set; }
